+++ clx_invoice_reports/__manifest__.py
@@ -3,17 +3,19 @@
 # See LICENSE file for full copyright & licensing details.
 {
     'name': 'CLX Reports',
-    'version': '13.3.0.0.7',
+    'version': '13.3.0.0.8',
     'summary': 'CLX Reports',
     'sequence': 1,
     'description': """ CLX Reports """,
     'category': '',
     'author': 'CLx Media',
     'website': 'https://conversionlogix.com/',
-    'depends': ['account'],
+    'depends': ['account',
+                'clx_invoice_policy'
+                ],
     'data': [
-        # 'views/report_templates.xml',
-        # 'views/report_invoice_document.xml',
+        'views/report_templates.xml',
+        'views/report_invoice_document.xml',
         'views/res_company_views.xml'
 
     ],